restructured game and moved to vscode

+++ engine/graphics/sprite.cpp
@@ -1,13 +1,15 @@
 #include "engine/graphics/sprite.h"
 #include <iostream>
 
-namespace graphics
-{
+namespace graphics {
 
-Sprite::Sprite(Texture *pTexture, float pX, float pY, int pRows, int pCols,
-        int pSpriteNum) :
-        GameObject(pTexture, pX, pY), rows(pRows), cols(pCols), spriteNum(
-                pSpriteNum)
+Sprite::Sprite(Texture* pTexture, float pX, float pY, int pRows, int pCols,
+    int pSpriteNum)
+    : GameObject(pTexture, pX, pY)
+    , rows(pRows)
+    , cols(pCols)
+    , spriteNum(
+          pSpriteNum)
 {
     sprite_width = pTexture->getWidth() / pCols;
     sprite_height = pTexture->getHeight() / pRows;
@@ -16,7 +18,7 @@ Sprite::Sprite(Texture *pTexture, float pX, float pY, int pRows, int pCols,
     lastTime = 0;
     animActive = false;
     moveWithCamera = true;
-    collisionArea = {false,false,false,false};
+    collisionArea = { false, false, false, false };
 }
 
 Sprite::~Sprite()
@@ -24,11 +26,9 @@ Sprite::~Sprite()
     //dtor
 }
 
-void Sprite::render(core::Renderer *pRender)
+void Sprite::render(core::Renderer* pRender)
 {
     graphics::Rect srcRect, destRect;
-
-
 
     if (sprite_width == 0) {
         sprite_width = getTexture()->getWidth() / cols;
@@ -36,25 +36,24 @@ void Sprite::render(core::Renderer *pRender)
     }
     if (animActive)
         lastTime += pRender->getTimeDelta();
-    if (lastTime >= 200)
-    {
+    if (lastTime >= 200) {
         lastTime = 0;
-        animFrame += 1.0;
+        animFrame += 1.0f;
         if (animFrame > 2)
             animFrame = 0;
     }
 
     int spriteRow = spriteNum / (rows / 2);
     int spriteCol = (spriteNum - (spriteRow * (cols / 3)));
-    int dir = (int) direction;
+    int dir = (int)direction;
 
     int dir_x = ((spriteCol * 3) + abs(animFrame)) * sprite_width;
     int dir_y = sprite_height * (dir + spriteRow * 4);
-    double x = getX()
-            - ((moveWithCamera) ? pRender->getMainCamera()->getX() : 0);
-    double y = getY()
-            - ((moveWithCamera) ? pRender->getMainCamera()->getY() : 0);
-    double zoom = pRender->getZoomFactor();
+    float x = getX()
+        - ((moveWithCamera) ? pRender->getMainCamera()->getX() : 0);
+    float y = getY()
+        - ((moveWithCamera) ? pRender->getMainCamera()->getY() : 0);
+    float zoom = pRender->getZoomFactor();
 
     srcRect.x = dir_x;
     srcRect.y = dir_y;
@@ -67,7 +66,6 @@ void Sprite::render(core::Renderer *pRender)
     destRect.height = sprite_height * zoom;
     if (x >= 0 && y >= 0)
         this->getTexture()->render(pRender, srcRect, destRect);
-
 }
 void Sprite::startAnimation()
 {
@@ -96,11 +94,13 @@ Rect Sprite::collisionRect()
 
 } // namespace graphics
 
-void graphics::Sprite::update(double deltaTime) {
+void graphics::Sprite::update(double deltaTime)
+{
 }
 
-graphics::Sprite::Sprite(const Sprite& pOriginal) :
-        GameObject(pOriginal) {
+graphics::Sprite::Sprite(const Sprite& pOriginal)
+    : GameObject(pOriginal)
+{
     rows = pOriginal.rows;
     cols = pOriginal.cols;
 
@@ -111,8 +111,7 @@ graphics::Sprite::Sprite(const Sprite& pOriginal) :
     lastTime = 0;
     animActive = false;
     moveWithCamera = true;
-    collisionArea = {false,false,false,false};
+    collisionArea = { false, false, false, false };
     spriteNum = pOriginal.spriteNum;
     std::cout << "sprite was copied" << std::endl;
-
 }
